refactor(analysis): split AnalysisPanel.tsx to stay under 400-line limit
Extract WarningBanner/AnalysisPending/AnalysisEmpty to AnalysisStates.tsx,
pure helpers to analysisPanelUtils.ts, and all computed state/filtering to
useAnalysisState.ts hook. AnalysisPanel.tsx drops from 715 to 362 lines.

The hook also fixes a runtime bug introduced in the prior commit:
filteredDecisionsVisualizationData was placed before effectiveModels in the
render function, causing a temporal dead zone ReferenceError.
